requirements.txt
ImageHash==4.0
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
PyWavelets==1.0.3
scipy==1.3.1
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
